+++ library/src/main/java/org/apache/apex/malhar/lib/dimensions/aggregator/AggregatorCumSum.java
@@ -16,10 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.datatorrent.lib.dimensions.aggregator;
+package org.apache.apex.malhar.lib.dimensions.aggregator;
 
 import java.util.List;
 import java.util.Map;
+
+import org.apache.apex.malhar.lib.dimensions.DimensionsEvent.Aggregate;
+import org.apache.apex.malhar.lib.dimensions.DimensionsEvent.InputEvent;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -33,8 +36,6 @@ import com.datatorrent.lib.appdata.gpo.SerdeListGPOMutable;
 import com.datatorrent.lib.appdata.gpo.SerdeObjectPayloadFix;
 import com.datatorrent.lib.appdata.schemas.FieldsDescriptor;
 import com.datatorrent.lib.appdata.schemas.Type;
-import com.datatorrent.lib.dimensions.DimensionsEvent.Aggregate;
-import com.datatorrent.lib.dimensions.DimensionsEvent.InputEvent;
 
 @Name("CUM_SUM")
 /**
