+++ library/src/main/java/org/apache/apex/malhar/lib/dimensions/aggregator/AggregatorIncrementalType.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.datatorrent.lib.dimensions.aggregator;
+package org.apache.apex.malhar.lib.dimensions.aggregator;
 
 import java.util.Collections;
 import java.util.Map;
